lessons/lesson-02-faq-expert/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.32.0

# Document Processing (for indexer)
docling==2.65.0
docling-core

# Vector Database & Embeddings
psycopg2-binary
pgvector
openai
ollama

# Microsoft Agent Framework
agent-framework

# Environment and utilities
python-dotenv==1.0.1
